Split web progress storage modules
